+++ org.eclipse.jdt.core.tests.compiler/src/org/eclipse/jdt/core/tests/compiler/parser/SyntaxErrorTest.java
@@ -25,7 +25,7 @@ import org.eclipse.jdt.internal.compiler.problem.ProblemReporter;
 
 public class SyntaxErrorTest extends org.eclipse.jdt.core.tests.junit.extension.TestCase {
 	public static boolean optimizeStringLiterals = false;
-	public static int sourceLevel = CompilerOptions.JDK1_3; //$NON-NLS-1$
+	public static long sourceLevel = CompilerOptions.JDK1_3; //$NON-NLS-1$
 	
 	public SyntaxErrorTest(String testName){
 	super(testName);
