+++ org.eclipse.jdt.core/codeassist/org/eclipse/jdt/internal/codeassist/impl/AssistParser.java
@@ -76,7 +76,7 @@ public abstract class AssistParser extends Parser {
 	
 	protected boolean isFirst = false;
 
-public AssistParser(ProblemReporter problemReporter, int sourceLevel) {
+public AssistParser(ProblemReporter problemReporter, long sourceLevel) {
 	super(problemReporter, true, sourceLevel);
 }
 public abstract char[] assistIdentifier();
