+++ org.eclipse.jdt.core/codeassist/org/eclipse/jdt/internal/codeassist/select/SelectionParser.java
@@ -40,7 +40,7 @@ public class SelectionParser extends AssistParser {
 	public static final char[] SUPER = "super".toCharArray(); //$NON-NLS-1$
 	public static final char[] THIS = "this".toCharArray(); //$NON-NLS-1$
 	
-public SelectionParser(ProblemReporter problemReporter, int sourceLevel) {
+public SelectionParser(ProblemReporter problemReporter, long sourceLevel) {
 	super(problemReporter, sourceLevel);
 }
 public char[] assistIdentifier(){
