+++ org.eclipse.jdt.core/search/org/eclipse/jdt/internal/core/search/matching/MatchLocatorParser.java
@@ -99,7 +99,7 @@ public class LocalDeclarationVisitor extends AbstractSyntaxTreeVisitorAdapter {
 	
 }
 
-public MatchLocatorParser(ProblemReporter problemReporter, int sourceLevel) {
+public MatchLocatorParser(ProblemReporter problemReporter, long sourceLevel) {
 	super(problemReporter, true, sourceLevel);
 }
 protected void classInstanceCreation(boolean alwaysQualified) {
